+++ app/scripts/fhir/STU3/lforms-fhir-sdc-converter.js
@@ -20,7 +20,7 @@ function addSDCImportFns(ns) {
   self.fhirExtUrlAllowedUnits = "http://hl7.org/fhir/StructureDefinition/elementdefinition-allowedUnits";
   self.fhirExtUrlCodingInstructions = "http://hl7.org/fhir/StructureDefinition/questionnaire-displayCategory";
   self.fhirExtUrlOptionPrefix = "http://hl7.org/fhir/StructureDefinition/questionnaire-optionPrefix";
-  self.fhirExtUrlOptionScore = "http://hl7.org/fhir/StructureDefinition/questionnaire-optionScore";
+  self.fhirExtUrlOptionScore = "http://hl7.org/fhir/StructureDefinition/questionnaire-ordinalValue";
   self.fhirExtUrlRestrictionArray = [
     "http://hl7.org/fhir/StructureDefinition/minValue",
     "http://hl7.org/fhir/StructureDefinition/maxValue",
@@ -31,6 +31,8 @@ function addSDCImportFns(ns) {
   self.fhirExtUrlAnswerRepeats = "http://hl7.org/fhir/StructureDefinition/questionnaire-answerRepeats";
 
   self.fhirExtUrlExternallyDefined = "http://hl7.org/fhir/StructureDefinition/questionnaire-externallydefined";
+  self.argonautExtUrlExtensionScore = "http://fhir.org/guides/argonaut-questionnaire/StructureDefinition/extension-score";
+
   self.fhirExtUrlHidden = "http://hl7.org/fhir/StructureDefinition/questionnaire-hidden";
 
   /**
@@ -45,11 +47,12 @@ function addSDCImportFns(ns) {
     if(fhirData) {
       target = {};
       _processFormLevelFields(target, fhirData);
+      var containedVS = _extractContainedVS(fhirData);
 
       if(fhirData.item && fhirData.item.length > 0) {
         target.items = [];
         for( var i = 0; i < fhirData.item.length; i++) {
-          var item = self._processQuestionnaireItem(fhirData.item[i], fhirData);
+          var item = self._processQuestionnaireItem(fhirData.item[i], fhirData, containedVS);
           target.items.push(item);
         }
       }
@@ -80,16 +83,77 @@ function addSDCImportFns(ns) {
     }
   }
 
+  /**
+   * Extract contained VS (if any) from the given questionnaire resource object.
+   * @param questionnaire the FHIR questionnaire resource object
+   * @return when there are contained value sets, returns a hash from "#<ValueSet.id>" (the character "#"
+   *         followed by the ValueSet id) to the answers options object, which, in turn, is a hash with 4 entries:
+   *         - "answers" is the list of LF answers converted from the value set.
+   *         - "systems" is the list of code systems for each answer item; and
+   *         - "isSameCodeSystem" is a boolean flag, true IFF the code systems for all answers in the list are the same.
+   *         - "hasAnswerCodeSystems" is a boolean flag, true IFF at least one answer has code system.
+   *         returns undefined if no contained value set is present.
+   * @private
+   */
+  function _extractContainedVS(questionnaire) {
+    var answersVS;
+
+    if(questionnaire.contained && questionnaire.contained.length > 0) {
+      answersVS = {};
+      questionnaire.contained.forEach(vs => {
+        if(vs.resourceType === 'ValueSet' && vs.expansion && vs.expansion.contains && vs.expansion.contains.length > 0) {
+          var lfVS = {answers: [], systems:[]};
+          var theCodeSystem = '#placeholder#'; // the code system if all answers have the same code systems, or "null"
+          vs.expansion.contains.forEach(vsItem => {
+            var answer = {code: vsItem.code, text: vsItem.display};
+            var ordExt = LForms.Util.findObjectInArray(vsItem.extension, 'url',
+              "http://hl7.org/fhir/StructureDefinition/valueset-ordinalValue");
+            if(ordExt) {
+              answer.score = ordExt.valueDecimal;
+            }
+            lfVS.answers.push(answer);
+            lfVS.systems.push(vsItem.system);
+
+            if(theCodeSystem === '#placeholder#') {
+              theCodeSystem = vsItem.system;
+            }
+            else if(theCodeSystem !== vsItem.system) {
+              theCodeSystem = null;
+            }
+            if(vsItem.system) {
+              lfVS.hasAnswerCodeSystems = true;
+            }
+          });
+
+          // set a flag if all the answers have identical code system, e.g., for use in LF item.answerCodeSystem
+          if(theCodeSystem && theCodeSystem !== '#placeholder#' ) {
+            lfVS.isSameCodeSystem = true;
+          }
+
+          // support both id and url based lookup. STU3 reference is quite vague.
+          if(vs.id !== undefined) {
+            answersVS['#' + vs.id] = lfVS;
+          }
+          if(vs.url !== undefined) {
+            answersVS[vs.url] = lfVS;
+          }
+        }
+      });
+    }
+
+    return answersVS;
+  }
 
   /**
    * Process questionnaire item recursively
    *
    * @param qItem - item object as defined in FHIR Questionnaire.
    * @param qResource - The source object of FHIR  questionnaire resource to which the qItem belongs to.
+   * @param containedVS - contained ValueSet info, see _extractContainedVS() for data format details
    * @returns {{}} - Converted 'item' field object as defined by LForms definition.
    * @private
    */
-  self._processQuestionnaireItem = function (qItem, qResource) {
+  self._processQuestionnaireItem = function (qItem, qResource, containedVS) {
     var targetItem = {};
     targetItem.question = qItem.text;
     //A lot of parsing depends on data type. Extract it first.
@@ -106,14 +170,14 @@ function addSDCImportFns(ns) {
     _processUnitList(targetItem, qItem);
     self._processDefaultAnswer(targetItem, qItem);
     _processExternallyDefined(targetItem, qItem);
-    _processAnswers(targetItem, qItem);
+    _processAnswers(targetItem, qItem, containedVS);
     _processSkipLogic(targetItem, qItem, qResource);
     _processCalculatedValue(targetItem, qItem);
 
     if (Array.isArray(qItem.item)) {
       targetItem.items = [];
       for (var i=0; i < qItem.item.length; i++) {
-        var newItem = self._processQuestionnaireItem(qItem.item[i], qResource);
+        var newItem = self._processQuestionnaireItem(qItem.item[i], qResource, containedVS);
         targetItem.items.push(newItem);
       }
     }
@@ -228,27 +292,56 @@ function addSDCImportFns(ns) {
    *
    * @param lfItem {object} - LForms item object to assign answer list
    * @param qItem {object} - Questionnaire item object
+   * @param containedVS - contained ValueSet info, see _extractContainedVS() for data format details
    * @private
    */
-  function _processAnswers(lfItem, qItem) {
+  function _processAnswers(lfItem, qItem, containedVS) {
     if(qItem.option) {
       lfItem.answers = [];
       for(var i = 0; i < qItem.option.length; i++) {
         var answer = {};
-        var label = LForms.Util.findObjectInArray(qItem.option[i].extension, 'url', self.fhirExtUrlOptionPrefix);
+        var option = qItem.option[i];
+        var label = LForms.Util.findObjectInArray(option.extension, 'url', self.fhirExtUrlOptionPrefix);
         if(label) {
           answer.label = label.valueString;
         }
-        var score = LForms.Util.findObjectInArray(qItem.option[i].modifierExtension, 'url', self.fhirExtUrlOptionScore);
+        var score = LForms.Util.findObjectInArray(option.extension, 'url', self.fhirExtUrlOptionScore);
+        // Look for argonaut extension.
+        score = !score ? LForms.Util.findObjectInArray(option.extension, 'url', self.argonautExtUrlExtensionScore) : score;
         if(score) {
-          answer.score = score.valueInteger.toString();
+          answer.score = score.valueDecimal.toString();
         }
-        answer.code = qItem.option[i].valueCoding.code;
-        answer.text = qItem.option[i].valueCoding.display;
+        var optionKey = Object.keys(option).filter(function(key) {return (key.indexOf('value') === 0);});
+        if(optionKey && optionKey.length > 0) {
+          if(optionKey[0] === 'valueCoding') { // Only one value[x] is expected
+            if(option[optionKey[0]].code    !== undefined) answer.code = option[optionKey[0]].code;
+            if(option[optionKey[0]].display !== undefined) answer.text = option[optionKey[0]].display;
+            //Lforms has answer code system at item level, expects all options to have one code system!
+            if(option[optionKey[0]].system  !== undefined) lfItem.answerCodeSystem = option[optionKey[0]].system;
+          }
+          else {
+            answer.text = option[optionKey[0]].toString();
+          }
+        }
+
         lfItem.answers.push(answer);
       }
     }
+    else if(qItem.options && containedVS) {
+      var vs = containedVS[qItem.options.reference];
+      if(vs) {
+        lfItem.answers = vs.answers;
+        if(vs.isSameCodeSystem) {
+          lfItem.answerCodeSystem = _toLfCodeSystem(vs.systems[0]);
+        }
+        else if(vs.hasAnswerCodeSystems) {
+          console.log('WARNING: unable to handle different answer code systems within a question (ignored): %s',
+                      vs.systems.join(', '));
+        }
+      }
+    }
   }
+
 
 
   /**
@@ -392,6 +485,24 @@ function addSDCImportFns(ns) {
 
 
   /**
+   * Convert the given code system to LForms internal code system. Currently
+   * only converts 'http://loinc.org' to 'LOINC' and returns all other input as is.
+   * @param codeSystem
+   * @private
+   */
+  function _toLfCodeSystem(codeSystem) {
+    var ret = codeSystem;
+    switch(codeSystem) {
+      case 'http://loinc.org':
+        ret = 'LOINC';
+        break;
+    }
+
+    return ret;
+  }
+
+
+  /**
    * Get an object with code and code system
    *
    * @param questionnaireItemOrResource {object} - question
@@ -402,32 +513,19 @@ function addSDCImportFns(ns) {
     if(questionnaireItemOrResource &&
          Array.isArray(questionnaireItemOrResource.code) &&
          questionnaireItemOrResource.code.length) {
-      code = {};
-      switch(questionnaireItemOrResource.code[0].system) {
-        case 'http://loinc.org':
-          code.system = 'LOINC';
-          break;
-        default:
-          code.system = questionnaireItemOrResource.code[0].system;
-          break;
-      }
-
-      code.code = questionnaireItemOrResource.code[0].code;
+      code = {
+        code: questionnaireItemOrResource.code[0].code,
+        system: _toLfCodeSystem(questionnaireItemOrResource.code[0].system)
+      };
     }
     // If code is missing look for identifier.
     else if(questionnaireItemOrResource &&
       Array.isArray(questionnaireItemOrResource.identifier) &&
       questionnaireItemOrResource.identifier.length) {
-      code = {};
-      switch(questionnaireItemOrResource.identifier[0].system) {
-        case 'http://loinc.org':
-          code.system = 'LOINC';
-          break;
-        default:
-          code.system = questionnaireItemOrResource.identifier[0].system;
-          break;
-      }
-      code.code = questionnaireItemOrResource.identifier[0].value;
+      code = {
+        code: questionnaireItemOrResource.identifier[0].value,
+        system: _toLfCodeSystem(questionnaireItemOrResource.identifier[0].system)
+      };
     }
 
     return code;